Added comments, independence of line movement ...
Edits for release 2.0. 
Based on the results of initial testing, no errors were detected...

+++ RandomSymbol.cc
@@ -13,7 +13,7 @@ RandomSymbol::RandomSymbol(bool isPolychromy)
 	}
 }
 
-void RandomSymbol::printRandomSymbol(int yCoordinate, int xCoordinate)
+void RandomSymbol::printRandomSymbol(short yCoordinate, short xCoordinate)
 {
 	mvaddch(yCoordinate, xCoordinate, symbol | attribute | MY_ATTRIBUTES);
 }
